Publish initial grain integration CSV
The reason for using the @topocount scope is that no active
devs have access to the @sourcecred scope at the moment.

This will serve as a temporary solution for getting our core >v0.9.2
releases working again, so we can begin utilizing that functionality.

Once we have access to @sourcecred the package in the @topocount scope
will be deprecated in favor of the org scope.

In order to prevent this issue from occuring again, the release process
will be modified to test a local production build. Currently, local
builds with dev-dependencies installed are tested, which caused this
oversight.

Test plan:
This test plan relies on a global installation of lerna:
```sh
yarn global add lerna
```

1. `yarn clean && yarn && lerna run build` Create a fresh local build to
   be used with `scDev`
2. `yarn clean` remove all dependencies, we need to replicate a prod
   install next
3. `lerna bootstrap --no-hoist '*' -- --production` Install dependencies
   like they would be installed in an instance. the flow-mono.sh script
   execution should error. This is okay.
4. In a local sourcecred/cred instance, run  through the manual test
   script here: https://roamresearch.com/#/app/SourceCred/page/nHxYhbtca

All executions with `scdev` should executed without error.